+++ modules/app-desktop/src/main/kotlin/org/archivekeep/app/desktop/ui/dialogs/storages/AbstractStorageDialog.kt
@@ -98,7 +98,9 @@ abstract class AbstractStorageDialog<T_State : AbstractStorageDialog.IState, T_V
                         },
                         content = {},
                         bottomContent = {
-                            DialogDismissButton("Dismiss", onClick = onClose)
+                            DialogButtonContainer {
+                                DialogDismissButton("Dismiss", onClick = onClose)
+                            }
                         },
                     )
                 }